requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.5
WTForms==3.0.1
email-validator==2.1.0.post1
requests==2.31.0
python-dotenv==1.0.0
python-telegram-bot==20.3
gunicorn==21.2.0
bcrypt==4.1.2
schedule==1.2.0
